UDD - ScoreStore
ScoreManager 싱글턴패턴 적용
다른 class에서 할당이 아닌 ScoreManager 호출로 변경
loding -> loading

+++ Assets/Scripts/Score/ScoreData.cs
@@ -8,5 +8,5 @@ public class ScoreData
     public int waveCount;
     public int enemyKills;
     public int LV;
-    public Dictionary<string, float> weaponDamages = new Dictionary<string, float>();
+    public Dictionary<string, int> weaponDamages = new Dictionary<string, int>();
 }
